+++ src/main/java/frc/robot/commands/OneBar/OnebarDown.java
@@ -5,17 +5,13 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj.AnalogPotentiometer;
 import frc.robot.subsystems.OnebarSubsystem;
-import frc.robot.Constants;
 
-public class OnebarIn extends CommandBase {
-  /** Creates a new OnebarIn. */
+public class OnebarDown extends CommandBase {
   private final OnebarSubsystem onebarsubsystem;
-
-  public OnebarIn(OnebarSubsystem subsystem) {
+  
+  /** Creates a new OneBarDown. */
+  public OnebarDown(OnebarSubsystem subsystem) {
     onebarsubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
@@ -23,13 +19,14 @@ public class OnebarIn extends CommandBase {
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
+  public void initialize(){
+    onebarsubsystem.InverseMotor();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    onebarsubsystem.armIn();
+    onebarsubsystem.armDown();
   }
 
   // Called once the command ends or is interrupted.
