+++ src/main/java/frc/robot/commands/OneBar/OnebarOut.java
@@ -6,29 +6,26 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.OnebarSubsystem;
-import edu.wpi.first.wpilibj.DutyCycleEncoder;
-import frc.robot.Constants;
 
-public class OnebarDown extends CommandBase {
-  private final OnebarSubsystem onebarsubsystem;
+public class OnebarOut extends CommandBase {
+  /** Creates a new OnebarOut. */
+  private final OnebarSubsystem onebarSubsystem;
   
-  /** Creates a new OneBarDown. */
-  public OnebarDown(OnebarSubsystem subsystem) {
-    onebarsubsystem = subsystem;
+  public OnebarOut(OnebarSubsystem subsystem) {
+    onebarSubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize(){
-    onebarsubsystem.InverseMotor();
+  public void initialize() {
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    onebarsubsystem.armDown();
+    onebarSubsystem.armOut();
   }
 
   // Called once the command ends or is interrupted.
