+++ src/main/java/frc/robot/commands/OneBar/OnebarUp.java
@@ -6,8 +6,6 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.OnebarSubsystem;
-import edu.wpi.first.wpilibj.DutyCycleEncoder;
-import frc.robot.Constants;
 
 public class OnebarUp extends CommandBase {
   private final OnebarSubsystem onebarsubsystem;
